+++ app/controllers/topics_controller.rb
@@ -1,7 +1,8 @@
 class TopicsController < ApplicationController
     before_action :require_sign_in, except: [:index, :show]
-    before_action :authorize_user, except: [:index, :show]
-    
+    before_action :authorize_user, except: [:index, :show, :edit, :update]
+    before_action :authorize_user_for_update, only: [:edit, :update]
+    before_action :authorize_user_for_delete, only: [:destroy]
     def index
         @topics = Topic.all
     end
@@ -43,13 +44,16 @@ class TopicsController < ApplicationController
  
     def destroy
         @topic = Topic.find(params[:id])
-
-        if @topic.destroy
-            flash[:notice] = "\"#{@topic.name}\" was deleted successfully."
+        if current_user.admin?
+            if @topic.destroy
+                flash[:notice] = "\"#{@topic.name}\" was deleted successfully."
+                redirect_to action: :index
+            else
+                flash.now[:alert] = "There was an error deleting the topic."
+                render :show
+            end
+        elsif current_user.moderator?
             redirect_to action: :index
-        else
-            flash.now[:alert] = "There was an error deleting the topic."
-            render :show
         end
     end
     private
@@ -57,6 +61,18 @@ class TopicsController < ApplicationController
             params.require(:topic).permit(:name, :description, :public)
         end
 
+        def authorize_user_for_update
+            unless current_user.moderator? || current_user.admin?
+                flash[:alert] = "You must be an admin to do that."
+                redirect_to topics_path
+            end
+        end
+        def authorize_user_for_delete
+            unless current_user.admin?
+                flash[:alert] = "You must be an admin to do that."
+                redirect_to topics_path
+            end
+        end
         def authorize_user
             unless current_user.admin?
                 flash[:alert] = "You must be an admin to do that."
